requirements.txt
Flask==3.0.3
google-generativeai==0.8.2
requests==2.32.3
Pillow==10.4.0
flask-cors==5.0.0
numpy==2.1.2
opencv-python-headless==4.10.0.84
jsonschema==4.23.0
